Implement Marks feedback
